Refactor form item editor components for smooth transition
